br.fosge.engine.graphics (primitives)

+++ fosge-engine/src/main/java/br/fosge/engine/runtime/ecs/component/MeshComponent.java
@@ -1,6 +1,5 @@
 package br.fosge.engine.runtime.ecs.component;
 
-import br.fosge.RT;
 import br.fosge.commons.Logger;
 import br.fosge.commons.Strings;
 import br.fosge.commons.Tuple;
@@ -15,12 +14,14 @@ import br.fosge.engine.runtime.ecs.ComponentType;
 import java.util.ArrayList;
 import java.util.List;
 
+import static br.fosge.RT.Graphics.textureUnitLimit;
+
 public final class MeshComponent extends Component {
     public static final ComponentType type = ComponentType.MESH_COMPONENT;
 
     public Shader shader;
     public Geometry geometry;
-    public final List<Texture> textures = new ArrayList<>(RT.Graphics.textureUnitLimit);
+    public final List<Texture> textures = new ArrayList<>(textureUnitLimit);
 
     private MeshComponent() {}
 
@@ -36,7 +37,7 @@ public final class MeshComponent extends Component {
 
         Logger.trace("Attaching shader: %s", instance.shader);
 
-        for (int i = 0; i < RT.Graphics.textureUnitLimit; i++) {
+        for (int i = 0; i < textureUnitLimit; i++) {
             final var candidate = find("texture." + i + ".asset", properties);
             // Expect bind unit to be contiguous
             if (candidate == null) { break; }
@@ -52,31 +53,39 @@ public final class MeshComponent extends Component {
             instance.textures.add(texture);
         }
 
-        final var layouts = new ArrayList<BufferLayout>();
-        for (final var tuple : properties) {
-            if (tuple.name().startsWith("geometry.layout")) {
-                layouts.add(new BufferLayout(
-                        tuple.name().substring(tuple.name().lastIndexOf('.') + 1),
-                        BufferType.valueOf(tuple.value())
-                ));
+        final var primitive = find("geometry.primitive", properties);
+        if (instance.geometry == null && primitive == null) {
+            final var layouts = new ArrayList<BufferLayout>();
+            for (final var tuple : properties) {
+                if (tuple.name().startsWith("geometry.layout")) {
+                    layouts.add(new BufferLayout(
+                            tuple.name().substring(tuple.name().lastIndexOf('.') + 1),
+                            BufferType.valueOf(tuple.value())
+                    ));
+                }
             }
+
+            instance.geometry = Resources.geometry(new GeometrySpec(
+                    DrawMode.valueOf(find("geometry.mode", properties)),
+                    DataType.valueOf(find("geometry.type", properties)),
+                    layouts.toArray(new BufferLayout[0])
+            ));
+
+            if (instance.geometry == null) {
+                Logger.error("Failed to configure texture");
+                instance.terminate();
+                return null;
+            }
+
+            Logger.trace("Attaching geometry: %s", instance.geometry);
+            instance.geometry.elements(Strings.ints(find("geometry.elements", properties)));
+            instance.geometry.vertices(Strings.floats(find("geometry.vertices", properties)));
         }
 
-        instance.geometry = Resources.geometry(new GeometrySpec(
-                        DrawMode.valueOf(find("geometry.mode", properties)),
-                        DataType.valueOf(find("geometry.type", properties)),
-                        layouts.toArray(new BufferLayout[0])
-        ));
 
-        if (instance.geometry == null) {
-            Logger.error("Failed to configure texture");
-            instance.terminate();
-            return null;
+        if (instance.geometry == null && "quadrilateral".equals(primitive)) {
+            instance.geometry = Primitives.quadrilateral();
         }
-
-        Logger.trace("Attaching geometry: %s", instance.geometry);
-        instance.geometry.elements(Strings.ints(find("geometry.elements", properties)));
-        instance.geometry.vertices(Strings.floats(find("geometry.vertices", properties)));
 
         return instance;
     }
